# ToDOv4/api/openapi.yaml
openapi: 3.0.0
info:
  title: Simple Task Manager API
  description: This is a simple API
  contact:
    email: joao_nascimento6@hotmail.com
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  version: 1.0.0
servers:
- url: /
  description: SwaggerHub API Auto Mocking
paths:
  /tarefas/userId:
    get:
      tags:
      - Tarefa
      summary: Lista todas as tarefas do user logado.
      description: Lista todas as tarefas do user logado.
      operationId: listarTodasTarefas
      parameters:
      - name: userId
        in: query
        description: Insere um ID de User no formato uuid
        required: true
        style: form
        explode: true
        schema:
          type: string
          format: uuid
          example: 550e8400-e29b-41d4-a726-446655440044
      responses:
        "200":
          description: Lista de todas as tarefas deste user bem solicitada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/listaDeTarefas'
        "204":
          description: O Id que procura não tem tarefas associadas.
        "400":
          description: O id do utlizador tem de coresponder a um formato uuid
        "401":
          description: Unauthorized - Forneça credenciais válidas.
        "500":
          description: Erro interno do servidor.
      security:
      - bearerAuth: []
      x-swagger-router-controller: Tarefa
  /tarefas/userId/prioridade:
    get:
      tags:
      - Tarefa
      summary: Lista todas as tarefas por prioridade
      description: Lista todas as tarefas por prioridade
      operationId: listarPrioridadeTarefas
      parameters:
      - name: userId
        in: query
        description: Insere um ID de user no formato uuid
        required: true
        style: form
        explode: true
        schema:
          type: string
          format: uuid
          example: 550e8400-e29b-41d4-a726-446655440044
      - name: prioridade
        in: query
        description: Seleciona o grau de prioridade da tarefa
        required: true
        style: form
        explode: true
        schema:
          type: integer
          enum:
          - 20
          - 19
          - 18
          - 17
          - 16
          - 15
          - 14
          - 13
          - 12
          - 11
          - 10
          - 9
          - 8
          - 7
          - 6
          - 5
          - 4
          - 3
          - 2
          - 1
      responses:
        "200":
          description: Resposta bem-sucedida. Lista de tarefas com a prioridade X
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/listaDeTarefasPorPrioridade'
        "204":
          description: O ID que procura não tem tarefas com este grau de   prioridade.
        "400":
          description: Requisição inválida. Verifique os parâmetros de entrada.
        "401":
          description: Unauthorized - Forneça credenciais válidas.
        "500":
          description: Erro interno do servidor.
      security:
      - bearerAuth: []
      x-swagger-router-controller: Tarefa
  /tarefas/userId/dataVencimento:
    get:
      tags:
      - Tarefa
      summary: Lista todas as tarefas por data de vencimento
      description: Lista todas as tarefas agrupadas por data de vencimento
      operationId: listarDataTarefas
      parameters:
      - name: userId
        in: query
        description: Insira um ID de usuário no formato uuid
        required: true
        style: form
        explode: true
        schema:
          type: string
          format: uuid
          example: 550e8400-e29b-41d4-a726-446655440044
      - name: dataVencimento
        in: query
        description: Data de vencimento no formato YYYY-MM-DD
        required: true
        style: form
        explode: true
        schema:
          pattern: "^\\d{4}-\\d{2}-\\d{2}T\\d{2}:\\d{2}:\\d{2}(.\\d{1,9})?Z?$"
          type: string
          format: date-time
          example: 2023-12-31T23:59:59.999999999Z
      responses:
        "200":
          description: Resposta bem-sucedida. Lista de tarefas agrupadas por   data
            de vencimento.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/listaDeTarefasDataVencimento'
        "204":
          description: O ID fornecido não tem tarefas associadas nesta data.
        "400":
          description: Requisição inválida. Verifique os parâmetros de entrada.
        "401":
          description: Unauthorized - Forneça credenciais válidas.
        "500":
          description: Erro interno do servidor.
      security:
      - bearerAuth: []
      x-swagger-router-controller: Tarefa
  /tarefas/userId/estado:
    get:
      tags:
      - Tarefa
      summary: Lista todas as tarefas por estado
      description: Lista todas as tarefas por estado
      operationId: listarEstadoTarefas
      parameters:
      - name: userId
        in: query
        description: Insere um ID de User no formato uuid
        required: true
        style: form
        explode: true
        schema:
          type: string
          format: uuid
          example: 550e8400-e29b-41d4-a726-446655440044
      - name: estado
        in: query
        description: "Estado da tarefa (Concluido, hoje, em progresso, futuro)."
        required: true
        style: form
        explode: true
        schema:
          type: string
          enum:
          - Concluido
          - Hoje
          - em progresso
          - futuro
      responses:
        "200":
          description: Resposta bem-sucedida
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/listaDeTarefasEstado'
        "204":
          description: O ID que procura não tem tarefas com esse estado.
        "400":
          description: Requisição inválida. Verifique os parâmetros de entrada.
        "401":
          description: Unauthorized - Forneça credenciais válidas.
        "500":
          description: Erro interno do servidor.
      security:
      - bearerAuth: []
      x-swagger-router-controller: Tarefa
    patch:
      tags:
      - Tarefa
      summary: Atualizar o estado de uma tarefa
      description: Atualiza o estado de uma tarefa com base no ID do usuário e no   estado
        fornecido.
      operationId: updateEstadoTarefas
      parameters:
      - name: userId
        in: query
        description: Insere um ID de User no formato uuid
        required: true
        style: form
        explode: true
        schema:
          type: string
          format: uuid
          example: 550e8400-e29b-41d4-a726-446655440044
      - name: estado
        in: query
        required: true
        style: form
        explode: true
        schema:
          type: string
          enum:
          - Concluido
          - Hoje
          - Em Progresso
          - Vencido
      requestBody:
        description: Objeto contendo as atualizações do estado da tarefa
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/userId_estado_body'
        required: true
      responses:
        "201":
          description: Tarefa atualizada com sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/tarefa'
        "400":
          description: Requisição inválida. Verifique os parâmetros de entrada.
        "401":
          description: Unauthorized - Forneça credenciais válidas.
        "500":
          description: Erro interno do servidor.
      security:
      - bearerAuth: []
      x-swagger-router-controller: Tarefa
  /tarefaspadrao:
    get:
      tags:
      - TarefaPadrao
      summary: listar tarefas padrao
      description: Devolve todas as tarefas padrao
      operationId: listarTarefaspadrao
      responses:
        "200":
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/listaTarefasPadrao'
        "401":
          description: Unauthorized - Forneça credenciais válidas.
        "404":
          description: Nao foram encontradas Tarefas Padrão.
        "500":
          description: Erro interno do servidor.
      security:
      - bearerAuth: []
      x-swagger-router-controller: TarefaPadrao
    post:
      tags:
      - TarefaPadrao
      summary: Criar uma tarefa padrão
      description: Cria uma nova tarefa padrão com os parâmetros especificados.
      operationId: criarTarefaPadrao
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/tarefasPadrao'
        required: true
      responses:
        "201":
          description: Tarefa padrão criada com sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/tarefasPadrao'
        "400":
          description: Requisição inválida - Verifique os parâmetros fornecidos.
        "401":
          description: Unauthorized - Forneça credenciais válidas.
        "500":
          description: Erro interno do servidor.
      security:
      - bearerAuth: []
      x-swagger-router-controller: TarefaPadrao
  /tarefaspadrao/tarefaId:
    delete:
      tags:
      - TarefaPadrao
      summary: Eliminar uma tarefa padrao por ID
      description: Elimina uma tarefa registada pelo seu ID
      operationId: deleteTarefaspadrao
      parameters:
      - name: tarefaId
        in: query
        description: Insere um ID de tarefa no formato uuid
        required: true
        style: form
        explode: true
        schema:
          type: string
          example: 550e8400-e29b-41d4-a726-446655440044
      responses:
        "204":
          description: Tarefa excluída com sucesso.
        "400":
          description: Insira o Id da tarefa correto.
        "401":
          description: Unauthorized - Forneça credenciais válidas.
        "404":
          description: Tarefa não encontrada.
        "500":
          description: Erro interno do servidor.
      security:
      - bearerAuth: []
      x-swagger-router-controller: TarefaPadrao
  /tarefaspadrao/tarefaId/dataVencimento:
    patch:
      tags:
      - TarefaPadrao
      summary: Atualizar a data de vencimento de uma tarefa por ID
      description: Atualiza a data de vencimento de uma tarefa registrada pelo seu
        ID
      operationId: updateDataTarefaspadrao
      parameters:
      - name: tarefaId
        in: query
        required: true
        style: form
        explode: true
        schema:
          type: string
          format: uuid
      requestBody:
        description: Objeto contendo a nova data de vencimento da tarefa
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/tarefaId_dataVencimento_body'
        required: true
      responses:
        "201":
          description: Tarefa atualizada com sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/tarefasPadrao'
        "401":
          description: Unauthorized - Forneça credenciais válidas.
        "404":
          description: Tarefa não encontrada.
        "500":
          description: Erro interno do servidor.
      security:
      - bearerAuth: []
      x-swagger-router-controller: TarefaPadrao
  /tarefaspadrao/userId/tarefaId:
    post:
      tags:
      - TarefaPadrao
      summary: Associar uma tarefa padrão a um usuário
      description: Cria uma associação entre um user e uma tarefa padrão com base
        em seus IDs.
      operationId: associaTarefapadraoUser
      parameters:
      - name: userId
        in: path
        description: Insere um ID de User no formato uuid
        required: true
        explode: true
        schema:
          type: string
          format: uuid
          example: 550e8400-e29b-41d4-a726-446655440044
      - name: tarefaId
        in: path
        description: Insere um ID de tarefa no formato uuid
        required: true
        explode: true
        schema:
          type: string
          example: 550e8400-e29b-41d4-a726-446655440044
      requestBody:
        description: Objeto contendo informações para associar a tarefa padrão ao
          usuário
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/tarefaPadraosUser'
        required: true
      responses:
        "200":
          description: Tarefa padrão associada ao usuário com sucesso.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/tarefaPadraosUser'
        "400":
          description: Requisição inválida. Verifique os parâmetros de entrada.
        "500":
          description: Erro interno do servidor.
      security:
      - bearerAuth: []
      x-swagger-router-controller: TarefaPadrao
...
  /tarefas:
    post:
      tags:
      - Tarefa
      summary: Criar uma nova tarefa
      description: Endpoint para criar uma nova tarefa.
      operationId: criarTarefas
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/tarefa'
        required: true
      responses:
        "201":
          description: Tarefa criada com sucesso.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/tarefa'
        "400":
          description: Requisição inválida. Verifique os parâmetros de entrada.
        "500":
          description: Erro interno do servidor.
      security:
      - bearerAuth: []
      x-swagger-router-controller: Tarefa
  /tarefas/{tarefaId}:
    get:
      tags:
      - Tarefa
      summary: Obter informações sobre uma tarefa
      description: Obter informações detalhadas sobre uma tarefa apartir do ID.
      operationId: listarIdTarefas
      parameters:
      - name: tarefaId
        in: path
        description: ID da tarefa que será atualizada
        required: true
        explode: true
        schema:
          type: string
          format: uuid
          example: 550e8400-e29b-41d4-a726-446655440044
      responses:
        "200":
          description: Informações detalhadas sobre a tarefa.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/tarefa'
        "400":
          description: Requisição inválida. Verifique os parâmetros de entrada.
        "401":
          description: Unauthorized - Forneça credenciais válidas.
        "404":
          description: Tarefa não encontrada.
        "500":
          description: Erro interno do servidor.
      security:
      - bearerAuth: []
      x-swagger-router-controller: Tarefa
    put:
      tags:
      - Tarefa
      summary: Atualizar uma tarefa existente
      description: Endpoint para atualizar uma tarefa existente.
      operationId: updateIdTarefas
      parameters:
      - name: tarefaId
        in: path
        required: true
        explode: true
        schema:
          type: string
          format: uuid
          example: 550e8400-e29b-41d4-a726-446655440044
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/tarefa'
        required: true
      responses:
        "201":
          description: Atualizar uma tarefa apartir do ID
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/tarefa'
        "400":
          description: Requisição inválida. Verifique os parâmetros de entrada.
        "401":
          description: Unauthorized - Forneça credenciais válidas.
        "404":
          description: Tarefa não encontrada.
        "500":
          description: Erro interno do servidor.
      security:
      - bearerAuth: []
      x-swagger-router-controller: Tarefa
    delete:
      tags:
      - Tarefa
      summary: Excluir uma tarefa existente
      description: Excluir uma tarefa existente apartir do ID.
      operationId: deleteTarefas
      parameters:
      - name: tarefaId
        in: path
        required: true
        explode: true
        schema:
          type: string
          format: uuid
          example: 550e8400-e29b-41d4-a726-446655440044
      responses:
        "200":
          description: Tarefa excluída com sucesso.
        "400":
          description: Insira o Id da tarefa correto.
        "401":
          description: Unauthorized - Forneça credenciais válidas.
        "404":
          description: Tarefa não encontrada.
        "500":
          description: Erro interno do servidor.
      security:
      - bearerAuth: []
      x-swagger-router-controller: Tarefa
components:
  schemas:
    tarefa:
      required:
      - dataVencimento
      - descrição
      - estado
      - prioridade
      - título
      - userId
      type: object
      properties:
        tarefaId:
          type: string
          format: uuid
        userId:
          type: string
          description: O ID do usuário associado à tarefa.
          format: uuid
        título:
          minLength: 1
          type: string
          description: O título da tarefa.
        descrição:
          type: string
          description: A descrição da tarefa.
        dataVencimento:
          pattern: "^\\d{4}-\\d{2}-\\d{2}T\\d{2}:\\d{2}:\\d{2}(.\\d{1,9})?Z?$"
          type: string
          description: A data de vencimento da tarefa (ISO 8601).
          format: date-time
        prioridade:
          maximum: 20
          minimum: 0
          type: integer
          description: A prioridade da tarefa (de 0 a 20).
        estado:
          type: string
          description: "O estado da tarefa (Concluido, em progresso ou futuro)."
          enum:
          - Concluido
          - Hoje
          - Agendado
      example:
        tarefaId: 550e8400-e29b-41d4-a716-446655440044
        userId: 550e8400-e29b-41d4-a716-446655440077
        título: Concluir Relatório Mensal
        descrição: Escrever um relatório sobre o desempenho mensal da equipe
        dataVencimento: 2023-12-31T23:59:59.999Z
        prioridade: 15
        estado: Concluido
    listaTarefasPadrao:
      type: array
      items:
        $ref: '#/components/schemas/listaTarefasPadrao_inner'
    tarefasPadrao:
      required:
      - dataVencimento
      - descricao
      - estado
      - prioridade
      - tarefaId
      - titulo
      type: object
      properties:
        dataVencimento:
          pattern: "^\\d{4}-\\d{2}-\\d{2}T\\d{2}:\\d{2}:\\d{2}(.\\d{1,9})?Z?$"
          type: string
          description: A data de vencimento da tarefa.
          format: date-time
        descricao:
          maxLength: 200
          minLength: 10
          type: string
          description: A descrição da tarefa.
        estado:
          type: string
          description: "O estado da tarefa (concluído, em progresso ou futuro)."
          enum:
          - Concluido
          - Hoje
          - Agendado
        prioridade:
          maximum: 20
          minimum: 0
          pattern: "^[0-9]+$"
          type: integer
          description: A prioridade da tarefa (de 0 a 20).
        tarefaId:
          pattern: "^[0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-5][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}$"
          type: string
          format: uuid
        titulo:
          maxLength: 30
          minLength: 4
          type: string
          description: O título da tarefa.
      example:
        tarefaId: 550e8400-e29b-41d4-a726-446655440044
        titulo: Terminar Relatorio
        descricao: Corrigir relatório sobre o desempenho mensal
        dataVencimento: 2023-12-31T23:59:59.999Z
        prioridade: 19
        estado: Hoje
    User:
      required:
      - email
      - password
      type: object
      properties:
        nome:
          type: string
        apelido:
          type: string
        email:
          pattern: "[^@ \\t\\r\\n]+@[^@ \\t\\r\\n]+[^@ \\t\\r\\n]+"
          type: string
        password:
          type: string
        contacto:
          pattern: "9[0-9]{2}[0-9]{3}[0-9]{3}$"
          type: string
        userId:
          pattern: "^[0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-5][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}$"
          type: string
          format: uuid
      example:
        password: password
        contacto: "912345678"
        apelido: apelido
        nome: nome
        email: email@hotmail.com
        userId: 550e8400-e29b-41d4-a716-446655440077
    listaDeTarefasDataVencimento:
      type: array
      example:
      - dataVencimento: 2023-12-31T23:59:59.999Z
        tarefas:
        - tarefaId: 550e8400-e29b-41d4-a716-446655440044
          userId: 550e8400-e29b-41d4-a716-446655440077
          título: Concluir Relatório Mensal
          descrição: Escrever um relatório sobre o desempenho mensal da equipe
          dataVencimento: 2023-12-31T23:59:59.999Z
          prioridade: 15
          estado: Concluido
        - tarefaId: 550e8400-e29b-41d4-a716-446655440045
          userId: 550e8400-e29b-41d4-a716-446655440078
          título: Reunião de Avaliação
          descrição: Realizar uma reunião para avaliar o progresso da equipe
          dataVencimento: 2023-12-31T23:59:59.999Z
          prioridade: 12
          estado: emProgresso
      - dataVencimento: 2023-12-30T23:59:59.999Z
        tarefas:
        - tarefaId: 550e8400-e29b-41d4-a716-446655440046
          userId: 550e8400-e29b-41d4-a716-446655440079
          título: Preparação para Apresentação
          descrição: Preparar material para a próxima apresentação da equipe
          dataVencimento: 2023-12-30T23:59:59.999Z
          prioridade: 18
          estado: futuro
      items:
        $ref: '#/components/schemas/listaDeTarefasDataVencimento_inner'
    listaDeTarefasEstado:
      type: array
      example:
      - estado: Concluido
        tarefas:
        - tarefaId: 550e8400-e29b-41d4-a716-446655440044
          userId: 550e8400-e29b-41d4-a716-446655440077
          título: Concluir Relatório Mensal
          descrição: Escrever um relatório sobre o desempenho mensal da equipe
          dataVencimento: 2023-12-31T23:59:59.999Z
          prioridade: 15
          estado: Concluido
      - estado: Hoje
        tarefas:
        - tarefaId: 550e8400-e29b-41d4-a716-446655440045
          userId: 550e8400-e29b-41d4-a716-446655440077
          título: Reunião de Equipe
          descrição: Reunião matinal para discutir os objetivos do dia
          dataVencimento: 2023-12-17T10:00:00.000Z
          prioridade: 10
          estado: Hoje
        - tarefaId: 550e8400-e29b-41d4-a716-446655440046
          userId: 550e8400-e29b-41d4-a716-446655440077
          título: Preparar Apresentação
          descrição: Criar slides para a apresentação da tarde
          dataVencimento: 2023-12-17T14:00:00.000Z
          prioridade: 12
          estado: Hoje
      - estado: Agendado
        tarefas:
        - tarefaId: 550e8400-e29b-41d4-a716-446655440047
          userId: 550e8400-e29b-41d4-a716-446655440077
          título: Projeto Futuro
          descrição: Iniciar planejamento para o próximo projeto
          dataVencimento: 2024-01-10T09:00:00.000Z
          prioridade: 18
          estado: Agendado
      items:
        $ref: '#/components/schemas/listaDeTarefasEstado_inner'
    listaDeTarefasPorPrioridade:
      type: array
      example:
      - prioridade: 15
        tarefas:
        - tarefaId: 550e8400-e29b-41d4-a716-446655440044
          userId: 550e8400-e29b-41d4-a716-446655440077
          título: Concluir Relatório Mensal
          descrição: Escrever um relatório sobre o desempenho mensal da   equipe
          dataVencimento: 2023-12-31T23:59:59.999Z
          prioridade: 15
          estado: Concluido
        - tarefaId: 550e8400-e29b-41d4-a716-446655440045
          userId: 550e8400-e29b-41d4-a716-446655440077
          título: Reunião com a Equipe
          descrição: Realizar uma reunião para discutir metas do próximo trimestre
          dataVencimento: 2023-12-15T15:00:00.000Z
          prioridade: 15
          estado: Hoje
      items:
        $ref: '#/components/schemas/listaDeTarefasPorPrioridade_inner'
    tarefaPadraosUser:
      required:
      - tarefasId
      - userId
      type: object
      properties:
        userId:
          $ref: '#/components/schemas/User'
        tarefasId:
          $ref: '#/components/schemas/tarefasPadrao'
      example:
        userId: 550e5400-e29b-41d4-a726-446655440044
        tarefasId: 550e8400-e29b-41d4-a726-446655440044
...
    listaDeTarefas:
      type: array
      example:
      - userId: 550e8400-e29b-41d4-a716-446655440077
        tarefas:
        - tarefaId: 550e8400-e29b-41d4-a716-446655440044
          userId: 550e8400-e29b-41d4-a716-446655440077
          título: Concluir Relatório Mensal
          descrição: Escrever um relatório sobre o desempenho mensal da equipe
          dataVencimento: 2023-12-31T23:59:59.999Z
          prioridade: 15
          estado: Concluido
        - tarefaId: 550e8400-e29b-41d4-a716-446655440045
          userId: 550e8400-e29b-41d4-a716-446655440077
          título: Reunião com a Equipe
          descrição: Realizar uma reunião para discutir metas do próximo trimestre
          dataVencimento: 2023-12-15T15:00:00.000Z
          prioridade: 12
          estado: Hoje
      items:
        $ref: '#/components/schemas/listaDeTarefas_inner'
    userId_estado_body:
      required:
      - novaEstado
      type: object
      properties:
        novaEstado:
          type: string
          enum:
          - Concluido
          - Hoje
          - Em Progresso
          - Vencido
    tarefaId_dataVencimento_body:
      type: object
      properties:
        dataVencimento:
          pattern: "^\\d{4}-\\d{2}-\\d{2}T\\d{2}:\\d{2}:\\d{2}(.\\d{1,9})?Z?$"
          type: string
          format: date-time
          example: 2023-12-31T23:59:59.999999999Z
...
    listaTarefasPadrao_inner:
      required:
      - dataVencimento
      - descrição
      - estado
      - prioridade
      - tarefaId
      - título
      type: object
      properties:
        dataVencimento:
          pattern: "^\\d{4}-\\d{2}-\\d{2}T\\d{2}:\\d{2}:\\d{2}(.\\d{1,9})?Z?$"
          type: string
          description: A data de vencimento da tarefa.
          format: date-time
        descrição:
          maxLength: 200
          minLength: 10
          type: string
          description: A descrição da tarefa.
        estado:
          type: string
          description: "O estado da tarefa (concluído, em progresso ou futuro)."
          enum:
          - Concluido
          - Hoje
          - Agendado
        prioridade:
          maximum: 20
          minimum: 0
          pattern: "^[0-9]+$"
          type: integer
          description: A prioridade da tarefa (de 0 a 20).
        tarefaId:
          pattern: "^[0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-5][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}$"
          type: string
          format: uuid
        título:
          maxLength: 30
          minLength: 4
          type: string
          description: O título da tarefa.
      example:
      - tarefaId: 550e8400-e29b-41d4-a726-446655440044
        título: Terminar Relatorio
        descrição: Corrigir relatório sobre o desempenho mensal
        dataVencimento: 2023-12-31T23:59:59.999Z
        prioridade: 19
        estado: Hoje
      - tarefaId: 550e8410-e29b-41d4-a726-446655440044
        título: Terminar Relatorio CG
        descrição: Corrigir relatório sobre o desempenho mensal
        dataVencimento: 2023-12-01T23:59:59.999Z
        prioridade: 11
        estado: Vencido
    listaDeTarefasDataVencimento_inner:
      required:
      - dataVencimento
      - tarefas
      type: object
      properties:
        dataVencimento:
          pattern: "^\\d{4}-\\d{2}-\\d{2}T\\d{2}:\\d{2}:\\d{2}(.\\d{1,9})?Z?$"
          type: string
          description: A data de vencimento da tarefa (ISO 8601).
          format: date-time
        tarefas:
          type: array
          items:
            $ref: '#/components/schemas/tarefa'
    listaDeTarefasEstado_inner:
      required:
      - estado
      - tarefas
      type: object
      properties:
        estado:
          type: string
          description: "O estado da tarefa (concluído, em progresso ou futuro)."
          enum:
          - Concluido
          - Hoje
          - Agendado
        tarefas:
          type: array
          items:
            $ref: '#/components/schemas/tarefa'
    listaDeTarefasPorPrioridade_inner:
      required:
      - prioridade
      - tarefas
      type: object
      properties:
        prioridade:
          maximum: 20
          minimum: 0
          pattern: "^[0-9]+$"
          type: integer
          description: A prioridade da tarefa (de 0 a 20).
        tarefas:
          type: array
          items:
            $ref: '#/components/schemas/tarefa'
    listaDeTarefas_inner:
      required:
      - tarefas
      - userId
      type: object
      properties:
        userId:
          type: string
          description: O ID do user associado às tarefas.
          format: uuid
        tarefas:
          type: array
          items:
            $ref: '#/components/schemas/tarefa'
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT